Architectural improvements
